+++ packages/Filters/src/Query/Filters/Fields/BaseFieldFilter.php
@@ -30,7 +30,7 @@ abstract class BaseFieldFilter extends FieldFilter
      *
      * @var string
      */
-    protected string $datetimeFormat = 'Y-m-d H:i:s';
+    protected string $databaseDatetimeFormat = 'Y-m-d H:i:s';
 
     /**
      * Map of operators (aliases) and corresponding database
@@ -53,6 +53,30 @@ abstract class BaseFieldFilter extends FieldFilter
         $this->operator = $this->operatorAliases()[$operator];
 
         return $this;
+    }
+
+    /**
+     * Set the datetime format used for queries
+     *
+     * @param string $format
+     *
+     * @return self
+     */
+    public function setDatabaseDatetimeFormat(string $format): static
+    {
+        $this->databaseDatetimeFormat = $format;
+
+        return $this;
+    }
+
+    /**
+     * Get the datetime format used for queries
+     *
+     * @return string
+     */
+    public function getDatabaseDatetimeFormat(): string
+    {
+        return $this->databaseDatetimeFormat;
     }
 
     /*****************************************************************
@@ -243,7 +267,7 @@ abstract class BaseFieldFilter extends FieldFilter
      * Builds "where [field] [operator] [datetime value]" constraint
      *
      * @param Builder|EloquentBuilder $query
-     * @param bool $utc  [optional]
+     * @param bool $utc  [optional] Converts datetime value to UTC if true
      *
      * @return Builder|EloquentBuilder
      */
@@ -287,10 +311,10 @@ abstract class BaseFieldFilter extends FieldFilter
 
         // Otherwise, for regular comparisons operators (<,>, <=, and >=)
         if ($this->logical() === FieldCriteria::OR) {
-            return $query->orWhere($field, $operator, $date->format($this->datetimeFormat));
+            return $query->orWhere($field, $operator, $date->format($this->getDatabaseDatetimeFormat()));
         }
 
-        return $query->where($field, $operator, $date->format($this->datetimeFormat));
+        return $query->where($field, $operator, $date->format($this->getDatabaseDatetimeFormat()));
     }
 
     /**
@@ -314,7 +338,7 @@ abstract class BaseFieldFilter extends FieldFilter
         int $offset = 1
     ): Builder|EloquentBuilder {
         // The general database datetime format to use.
-        $format = $this->datetimeFormat;
+        $format = $this->getDatabaseDatetimeFormat();
 
         // In case that no "seconds" precision is given, then ensure
         // that we increase the offset and adapt the format. This should